refactor: extract workspace worktree lifecycle
